refactor: config/config.ts => lib/consts.ts
