+++ src/components/modals/Settings/Settings.tsx
@@ -20,7 +20,7 @@ import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
 
 import CalendarFeed from "@/components/modals/Settings/CalendarFeed";
 import RippleBadge from "@/components/utils/RippleBadge";
-import { DEFAULT_REMINDER_HOURS } from "@/config/config";
+import { DEFAULT_REMINDER_HOURS } from "@/lib/consts";
 import { useIntegrationUser } from "@/lib/hooks/useIntegrationUser";
 import { usePreferences } from "@/lib/hooks/usePreferences";
 import { useUserConfig } from "@/lib/hooks/useUserConfig";
